finir parse double

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: ouamarko <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/09 18:20:05 by ouamarko          #+#    #+#             */
-/*   Updated: 2025/08/12 16:07:52 by ouamarko         ###   ########.fr       */
+/*   Updated: 2025/08/13 19:37:41 by ouamarko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 typedef struct s_list
 	{
-		int	content;
+		long	content;
 		struct s_list *next;
 	} t_list;
 
@@ -37,7 +37,7 @@ int	ft_isdigit(char s);
 t_list	*ft_lstnew(int content);
 void	ft_lstadd_front(t_list **lst, t_list *new);
 void	ft_lstadd_back(t_list **lst, t_list *new);
-int	ft_atoi(char *nptr);
+long	ft_atoi(char *nptr);
 char	**ft_split(char const *s, char c);
 int	ft_count_split(char **tab);
 int	ft_parse_solo_argument(char *tab);
